Move all info dialogs to their own package
