+++ android/app/src/main/kotlin/net/mullvad/mullvadvpn/compose/dialog/info/ApiAccessMethodInfoDialog.kt
@@ -1,9 +1,8 @@
-package net.mullvad.mullvadvpn.compose.dialog
+package net.mullvad.mullvadvpn.compose.dialog.info
 
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.tooling.preview.Preview
-import androidx.lifecycle.compose.dropUnlessResumed
 import com.ramcosta.composedestinations.annotation.Destination
 import com.ramcosta.composedestinations.annotation.RootGraph
 import com.ramcosta.composedestinations.navigation.DestinationsNavigator
@@ -15,20 +14,24 @@ import net.mullvad.mullvadvpn.lib.theme.AppTheme
 
 @Preview
 @Composable
-private fun PreviewLocalNetworkSharingInfoDialog() {
-    AppTheme { LocalNetworkSharingInfo(EmptyDestinationsNavigator) }
+private fun PreviewApiAccessMethodInfoDialog() {
+    AppTheme { ApiAccessMethodInfo(EmptyDestinationsNavigator) }
 }
 
 @Destination<RootGraph>(style = DestinationStyle.Dialog::class)
 @Composable
-fun LocalNetworkSharingInfo(navigator: DestinationsNavigator) {
+fun ApiAccessMethodInfo(navigator: DestinationsNavigator) {
     InfoDialog(
-        message = stringResource(id = R.string.local_network_sharing_info),
-        additionalInfo =
+        message =
             buildString {
-                appendLine(stringResource(id = R.string.local_network_sharing_additional_info))
-                appendLine(textResource(id = R.string.local_network_sharing_ip_ranges))
+                appendLine(stringResource(id = R.string.api_access_method_info_first_line))
+                appendLine()
+                appendLine(stringResource(id = R.string.api_access_method_info_second_line))
+                appendLine()
+                appendLine(textResource(id = R.string.api_access_method_info_third_line))
+                appendLine()
+                appendLine(textResource(id = R.string.api_access_method_info_fourth_line))
             },
-        onDismiss = dropUnlessResumed { navigator.navigateUp() },
+        onDismiss = navigator::navigateUp,
     )
 }
