+++ android/app/src/main/kotlin/net/mullvad/mullvadvpn/compose/dialog/info/CustomDnsInfoDialog.kt
@@ -1,4 +1,4 @@
-package net.mullvad.mullvadvpn.compose.dialog
+package net.mullvad.mullvadvpn.compose.dialog.info
 
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.res.stringResource
@@ -10,19 +10,18 @@ import com.ramcosta.composedestinations.navigation.DestinationsNavigator
 import com.ramcosta.composedestinations.navigation.EmptyDestinationsNavigator
 import com.ramcosta.composedestinations.spec.DestinationStyle
 import net.mullvad.mullvadvpn.R
-import net.mullvad.mullvadvpn.lib.theme.AppTheme
 
 @Preview
 @Composable
-private fun PreviewUdpOverTcpPortInfoDialog() {
-    AppTheme { UdpOverTcpPortInfo(EmptyDestinationsNavigator) }
+private fun PreviewCustomDnsInfoDialog() {
+    CustomDnsInfo(EmptyDestinationsNavigator)
 }
 
 @Destination<RootGraph>(style = DestinationStyle.Dialog::class)
 @Composable
-fun UdpOverTcpPortInfo(navigator: DestinationsNavigator) {
+fun CustomDnsInfo(navigator: DestinationsNavigator) {
     InfoDialog(
-        message = stringResource(id = R.string.udp_over_tcp_port_info),
+        message = stringResource(id = R.string.settings_changes_effect_warning_content_blocker),
         onDismiss = dropUnlessResumed { navigator.navigateUp() },
     )
 }
