+++ android/app/src/main/kotlin/net/mullvad/mullvadvpn/compose/dialog/info/LocalNetworkSharingInfoDialog.kt
@@ -1,4 +1,4 @@
-package net.mullvad.mullvadvpn.compose.dialog
+package net.mullvad.mullvadvpn.compose.dialog.info
 
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.res.stringResource
@@ -10,24 +10,24 @@ import com.ramcosta.composedestinations.navigation.DestinationsNavigator
 import com.ramcosta.composedestinations.navigation.EmptyDestinationsNavigator
 import com.ramcosta.composedestinations.spec.DestinationStyle
 import net.mullvad.mullvadvpn.R
+import net.mullvad.mullvadvpn.compose.component.textResource
+import net.mullvad.mullvadvpn.lib.theme.AppTheme
 
 @Preview
 @Composable
-private fun PreviewServerIpOverridesInfoDialog() {
-    ServerIpOverridesInfo(EmptyDestinationsNavigator)
+private fun PreviewLocalNetworkSharingInfoDialog() {
+    AppTheme { LocalNetworkSharingInfo(EmptyDestinationsNavigator) }
 }
 
 @Destination<RootGraph>(style = DestinationStyle.Dialog::class)
 @Composable
-fun ServerIpOverridesInfo(navigator: DestinationsNavigator) {
+fun LocalNetworkSharingInfo(navigator: DestinationsNavigator) {
     InfoDialog(
-        message =
+        message = stringResource(id = R.string.local_network_sharing_info),
+        additionalInfo =
             buildString {
-                appendLine(stringResource(id = R.string.server_ip_overrides_info_first_paragraph))
-                appendLine()
-                appendLine(stringResource(id = R.string.server_ip_overrides_info_second_paragraph))
-                appendLine()
-                append(stringResource(id = R.string.server_ip_overrides_info_third_paragraph))
+                appendLine(stringResource(id = R.string.local_network_sharing_additional_info))
+                appendLine(textResource(id = R.string.local_network_sharing_ip_ranges))
             },
         onDismiss = dropUnlessResumed { navigator.navigateUp() },
     )
