+++ android/app/src/main/kotlin/net/mullvad/mullvadvpn/compose/dialog/info/MalwareInfoDialog.kt
@@ -1,0 +1,28 @@
+package net.mullvad.mullvadvpn.compose.dialog.info
+
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.res.stringResource
+import androidx.compose.ui.tooling.preview.Preview
+import androidx.lifecycle.compose.dropUnlessResumed
+import com.ramcosta.composedestinations.annotation.Destination
+import com.ramcosta.composedestinations.annotation.RootGraph
+import com.ramcosta.composedestinations.navigation.DestinationsNavigator
+import com.ramcosta.composedestinations.navigation.EmptyDestinationsNavigator
+import com.ramcosta.composedestinations.spec.DestinationStyle
+import net.mullvad.mullvadvpn.R
+import net.mullvad.mullvadvpn.lib.theme.AppTheme
+
+@Preview
+@Composable
+private fun PreviewMalwareInfoDialog() {
+    AppTheme { MalwareInfo(EmptyDestinationsNavigator) }
+}
+
+@Destination<RootGraph>(style = DestinationStyle.Dialog::class)
+@Composable
+fun MalwareInfo(navigator: DestinationsNavigator) {
+    InfoDialog(
+        message = stringResource(id = R.string.malware_info),
+        onDismiss = dropUnlessResumed { navigator.navigateUp() },
+    )
+}
