+++ android/app/src/main/kotlin/net/mullvad/mullvadvpn/compose/dialog/info/WireguardPortInfoDialog.kt
@@ -1,4 +1,4 @@
-package net.mullvad.mullvadvpn.compose.dialog
+package net.mullvad.mullvadvpn.compose.dialog.info
 
 import android.os.Parcelable
 import androidx.compose.runtime.Composable
